refactor: move Editor units to src/Editor subdirectory
