+++ src/CMakeLists.txt
@@ -17,7 +17,7 @@ target_sources(kilo
         "${PROJECT_SOURCE_DIR}/src/TerminalState/TerminalState.hpp"
         "${PROJECT_SOURCE_DIR}/src/Cursor/Cursor.hpp"
         "${PROJECT_SOURCE_DIR}/src/Window/Window.hpp"
-        "${PROJECT_SOURCE_DIR}/includes/Editor.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Editor/Editor.hpp"
         "${PROJECT_SOURCE_DIR}/src/Utilities/Utilities.hpp"
         "${PROJECT_SOURCE_DIR}/includes/EditorConfig.hpp"
     PRIVATE
@@ -25,7 +25,7 @@ target_sources(kilo
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.cpp"
         "${PROJECT_SOURCE_DIR}/src/TerminalState/TerminalState.cpp"
         "${PROJECT_SOURCE_DIR}/src/Utilities/Utilities.cpp"
-        Editor.cpp
+        "${PROJECT_SOURCE_DIR}/src/Editor/Editor.cpp"
         EditorConfig.cpp
 )
